Update to 114 VM Tools Not Up to Date.ps1
Our environment uses productlocker so it can't use the result of "$_.Guest.toolsStatus -eq "toolsOld" as it wouldn't return the proper value.  Using a hard coded tools version is the best way to resolve this issue.  It will do a comparison of the tools version and $VMTools and if it is less than it will include it in the $Results.